refactor(backend): 🎨 shared cookie options
